Make the verifier accept the errno, and fix three tests the VM exposed
The action set compiled, passed go vet, passed every unit test, and was rejected
by the verifier on all four LSM programs: "the register R0 has unknown scalar
value should have been in [-4095, 0]". It would have shipped an agent that
attaches nothing and denies nothing while reporting itself healthy. This is what
make vm-test is for.

Three separate causes, each found only by reading the verifier's own trace.

The verifier does not propagate bounds through BPF_NEG. With e provably in
[1, 4095], the moment the compiler emits `r8 = -r8` the register becomes an
unknown scalar - the bound is true and the verifier will not take the code's
word for it. Two signed comparisons after the negation re-establish the range.
They can never fire; they exist to say so in a form the verifier accepts.

They also have to survive compilation. The first attempt used plain `if`
statements and the compiler deleted both, correctly, because it can prove them
dead. barrier_var is what keeps them in the object.

And the value has to be 64-bit. With `int`, the compiler zero-extends the
clamped result back to 64 bits - `r8 <<= 32; r8 >>= 32` - and the verifier then
sees 0x00000000fffff001, a large positive number, rather than -4095. The
register the kernel reads is 64 bits wide, so enforce_errno and enforce_apply
return long.

Three test fixes, two of which were the tests being wrong rather than the code:

The syscall monitor moved to tracepoint/raw_syscalls/sys_enter in the previous
commit and two VM tests still called AttachRawTracepoint, which fails with
"invalid program type TracePoint, expected RawTracepoint".

The audit test matched on /etc/os-release. bpf_d_path reports the RESOLVED path
and that file is a symlink into /usr/lib, so the match never fired and the
failure read as "the kernel emitted nothing" when the kernel had emitted exactly
the right event under a different name. It now creates its own file. This repo
already has a test named TestVMSeededSymlinkPathDoesNotMatch, which is a fair
comment on writing the same mistake again.

And it matched on path alone. The LSM hook is attached node-wide and the test
process creates the file it is about to read, so the first match was the test's
own write open from outside the governed cgroup entirely. Both read loops now
filter on cgroup id.

All 31 VM tests pass on kernel 6.8 with lsm=bpf: the five original programs, the
four LSM hooks under the new action encoding, syscall arguments decoded from a
kill(self, 0) whose arguments the test controls, a watched syscall reporting
three times where deduplication caps at one, commit_creds firing and reporting
only changes that gained privilege, Audit reporting an unlearned open and
allowing it and reporting it again on the second attempt, and a denial
configured with ENOENT reaching the caller as "No such file or directory".

+++ bpf/capability_monitor.c
@@ -135,7 +135,7 @@ int BPF_PROG(capable_check, const struct cred *cred, struct user_namespace *ns,
 			return 0; /* learned capability: permit */
 
 		__u32 flags = 0;
-		int ret = enforce_apply(spec, &flags, EV_DENIED, EV_KILLED);
+		long ret = enforce_apply(spec, &flags, EV_DENIED, EV_KILLED);
 
 		struct cap_event *e = bpf_ringbuf_reserve(&cap_events, sizeof(*e), 0);
 		if (e) {
